Extract Monitor interface
