+++ vm-monitor/src/main/java/com/epam/pipeline/vmmonitor/service/Monitor.java
@@ -1,0 +1,6 @@
+package com.epam.pipeline.vmmonitor.service;
+
+public interface Monitor {
+
+    void monitor();
+}
