+++ vm-monitor/src/main/java/com/epam/pipeline/vmmonitor/service/vm/VMMonitor.java
@@ -25,6 +25,7 @@ import com.epam.pipeline.entity.region.AbstractCloudRegion;
 import com.epam.pipeline.entity.region.CloudProvider;
 import com.epam.pipeline.exception.PipelineResponseException;
 import com.epam.pipeline.vmmonitor.model.vm.VirtualMachine;
+import com.epam.pipeline.vmmonitor.service.Monitor;
 import com.epam.pipeline.vmmonitor.service.pipeline.CloudPipelineAPIClient;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
@@ -52,7 +53,7 @@ import java.util.stream.Collectors;
 @Slf4j
 @Service
 @SuppressWarnings("PMD.AvoidCatchingGenericException")
-public class VMMonitor {
+public class VMMonitor implements Monitor {
 
     private static final String POOL_RUN_ID_PREFIX = "p-";
     private static final int SECONDS_IN_MINUTES = 60;
@@ -82,6 +83,7 @@ public class VMMonitor {
         this.vmMaxLiveMinutes = vmMaxLiveMinutes;
     }
 
+    @Override
     public void monitor() {
         try {
             ListUtils.emptyIfNull(apiClient.loadRegions())
